refactor(harness): extract built-in resource discovery owner
