+++ src/loushang/harness/resources/_loader_pipeline.py
@@ -8,11 +8,13 @@ from typing import TYPE_CHECKING
 from loushang.harness.diagnostics.types import DiagnosticDraft
 from loushang.harness.resources._loader_discovery import (
     _apply_resource_switches,
-    _discover_built_in_resources,
     _discover_external_package_resources,
     _discover_project_resources,
     _discover_temporary_resources,
     _discover_user_global_resources,
+)
+from loushang.harness.resources._loader_discovery_builtin import (
+    _discover_built_in_resources,
 )
 from loushang.harness.resources._loader_discovery_context import (
     _discover_context_descriptors,
